requirements.txt
asyncio
aiohttp
bs4
feedparser
pytz
pyyaml
validators
yattag